// src/generated/models/Dyn365bc_prodordercomponents_abk_prod_v1_0Model.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dyn365bc_prodordercomponents_abk_prod_v1_0dyn365bc_status = {
  200000000: 'Simulated',
  200000001: 'Planned',
  200000002: 'FirmPlanned',
  200000003: 'Released',
  200000004: 'Finished'
} as const;
export type Dyn365bc_prodordercomponents_abk_prod_v1_0dyn365bc_status = keyof typeof Dyn365bc_prodordercomponents_abk_prod_v1_0dyn365bc_status;

export interface Dyn365bc_prodordercomponents_abk_prod_v1_0Base {
  dyn365bc_bincode?: string;
  dyn365bc_companylookup_code?: string;
  "dyn365bc_companylookup_id@odata.bind": string;
  dyn365bc_description?: string;
  dyn365bc_duedate?: string;
  dyn365bc_expectedquantity?: number;
  dyn365bc_itemno?: string;
  dyn365bc_lineno?: number;
  dyn365bc_locationcode?: string;
  dyn365bc_prodordercomponent_abk_prod_v1_0id: string;
  dyn365bc_prodorderlineno?: number;
  dyn365bc_prodorderno?: string;
  dyn365bc_quantity?: number;
  dyn365bc_quantityper?: number;
  dyn365bc_remainingquantity?: number;
  dyn365bc_status?: Dyn365bc_prodordercomponents_abk_prod_v1_0dyn365bc_status;
  dyn365bc_unitofmeasurecode?: string;
  dyn365bc_variantcode?: string;
}

export interface Dyn365bc_prodordercomponents_abk_prod_v1_0 extends Dyn365bc_prodordercomponents_abk_prod_v1_0Base {
  dyn365bc_companylookup_idname?: string;
  dyn365bc_statusname?: string;
  dyn365bc_companylookup_id?: object;
  _dyn365bc_companylookup_id_value?: string;
}